+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [3.1.0] - 2020-04-09
 ### Added
+- Google Analytics v4 tracking script.
+- scripts folder containing scripts used in the application.
 - 500 static page.
 
 ### Changed
+- refactor: uses dynamic import for loading scripts without SSR.
 - `next@10.1.3`
 
 ### Fixed